Update Helm v2 stable chart repository URL
Since helm v2.17.0 the stable chart repositories were moved to a new URL, and since a few weeks ago the old repository is no longer available which is causing anyone using helm v2 with kubergrunt to fail.
Ideally everybody should be moving to helm v3, but this PR changes the URL to the new repository as a hotfix for those still using helm v2.

For context See:
- https://github.com/helm/helm/releases/tag/v2.17.0
- https://github.com/hashicorp/terraform-provider-helm/issues/649